requirements.txt
aiogram[fast]>=3.1.1
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.25


